% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{cycHumanBloodData}
\alias{cycHumanBloodData}
\title{cycHumanBloodData}
\format{A dataframe containing 439 columns (column 1 = transcript name,
  column 2 to 439 = samples from individuals at different time points and
  sleep conditions).}
\source{
Moller-Levet C. S., et al. (2013). Effects of insufficient sleep on
  circadian rhythmicity and expression amplitude of the human blood
  transcriptome. \emph{Proc Natl Acad Sci U S A}, \bold{110(12)},
  E1132--1141.
}
\usage{
cycHumanBloodData
}
\description{
This data set lists time-series profiles of 20 transcripts
  sampled from  multiple individuals under different sleep conditions.
}

