% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{cycHumanBloodDesign}
\alias{cycHumanBloodDesign}
\title{cycHumanBloodDesign}
\format{A dataframe containing 4 columns described as below:
  \tabular{rlll}{
    [,1] \tab sample_library  \tab character  \tab sample ID\cr
    [,2] \tab subject         \tab character  \tab individual ID\cr
    [,3] \tab group           \tab character  \tab sleep condition\cr
    [,4] \tab time_hoursawake \tab numeric    \tab hours after awake
  }}
\source{
Moller-Levet C. S., et al. (2013). Effects of insufficient sleep on
  circadian rhythmicity and expression amplitude of the human blood
  transcriptome. \emph{Proc Natl Acad Sci U S A}, \bold{110(12)},
  E1132--1141.
}
\usage{
cycHumanBloodDesign
}
\description{
This data set describes individual information, sleep
  condition and sampling time corresponding to each sample in
  'cycHumanBloodData'.
}

