% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{cycMouseLiverProtein}
\alias{cycMouseLiverProtein}
\title{cycMouseLiverProtein}
\format{A dataframe containing 49 columns(column 1 = protein name,
  column 2 to 49 = time points from CT0 to CT45 with three replicates
  at each time point).}
\source{
Robles M. S., Cox J., Mann M. (2014). In-vivo quantitative
  proteomics reveals a key contribution of post-transcriptional
  mechanisms to the circadian regulation of liver metabolism.
  \emph{PLoS Genet}, \bold{10(1)}, e1004047.
}
\usage{
cycMouseLiverProtein
}
\description{
This data set lists expression profiles of 20 circadian
  proteins with 3h-resolution covering two days.
}

