% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{cycMouseLiverRNA}
\alias{cycMouseLiverRNA}
\title{cycMouseLiverRNA}
\format{A dataframe containing 49 columns(column 1 = transcript name,
  column 2 to 49 = time points from CT18 to CT65).}
\source{
Hughes M. E., et al. (2009). Harmonics of circadian gene
  transcription in mammals. \emph{PLoS Genet}, \bold{5(4)}, e1000442.
}
\usage{
cycMouseLiverRNA
}
\description{
This data set lists expression profiles of 20 circadian
  transcripts with 1h-resolution covering two days.
}

