% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{cycSimu4h2d}
\alias{cycSimu4h2d}
\title{cycSimu4h2d}
\format{A dataframe containing 13 columns(column 1 = curve ID, column 2 to
  13 = time points from 0 to 44).}
\source{
Wu G., Zhu J., Yu J., Zhou L., Huang J. Z. and  Zhang Z. (2014).
  Evaluation of five methods for genome-wide circadian gene identification.
  \emph{Journal of Biological Rhythms}, \bold{29(4)}, 231--242.
}
\usage{
cycSimu4h2d
}
\description{
This data set lists 100 simulated profiles(periodic and
  non-periodic) with 4h-resolution covering two periods.
}

