% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{cycVignettesAMP}
\alias{cycVignettesAMP}
\title{cycVignettesAMP}
\format{A dataframe containing 71 columns described as below:
  \tabular{rlll}{
        [,1] \tab CycID                 \tab character  \tab transcript name\cr
        [,2] \tab ARS_pvalue            \tab numeric    \tab pvalue from ARS\cr
        [,3] \tab ARS_BH.Q              \tab numeric    \tab FDR from ARS\cr
        [,4] \tab ARS_period            \tab numeric    \tab period from ARS\cr
        [,5] \tab ARS_adjphase          \tab numeric    \tab adjusted phase from ARS\cr
        [,6] \tab ARS_amplitude         \tab numeric    \tab amplitude from ARS\cr
        [,7] \tab JTK_pvalue            \tab numeric    \tab pvalue from JTK\cr
        [,8] \tab JTK_BH.Q              \tab numeric    \tab FDR from JTK\cr
        [,9] \tab JTK_period            \tab numeric    \tab period from JTK\cr
       [,10] \tab JTK_adjphase          \tab numeric    \tab adjusted phase from JTK\cr
       [,11] \tab JTK_amplitude         \tab numeric    \tab amplitude from JTK\cr
       [,12] \tab LS_pvalue             \tab numeric    \tab pvalue from LS\cr
       [,13] \tab LS_BH.Q               \tab numeric    \tab FDR from JTK\cr
       [,14] \tab LS_period             \tab numeric    \tab period from LS\cr
       [,15] \tab LS_adjphase           \tab numeric    \tab adjusted phase from LS\cr
       [,16] \tab LS_amplitude          \tab numeric    \tab amplitude from LS\cr
       [,17] \tab meta2d_pvalue         \tab numeric    \tab integrated pvalue\cr
       [,18] \tab meta2d_BH.Q           \tab numeric    \tab FDR based on integrated pvalue\cr
       [,19] \tab meta2d_period         \tab numeric    \tab averaged period of three methods\cr
       [,20] \tab meta2d_phase          \tab numeric    \tab integrated phase\cr
       [,21] \tab meta2d_Base           \tab numeric    \tab baseline value given by meta2d\cr
       [,22] \tab meta2d_AMP            \tab numeric    \tab amplitude given by meta2d\cr
       [,23] \tab meta2d_rAMP           \tab numeric    \tab relative amplitude\cr
    [,24:71] \tab CT18 to CT65          \tab numeric    \tab sampling time point
  }}
\source{
Hughes M. E., et al. (2009). Harmonics of circadian gene
  transcription in mammals. \emph{PLoS Genet}, \bold{5(4)}, e1000442.
}
\usage{
cycVignettesAMP
}
\description{
This data set lists meta2d's analysis results of three
  circadian transcripts selected from the same source dataset used by
  cycMouseLiverRNA.
}

