% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsinfo.R
\docType{data}
\name{cycYeastCycle}
\alias{cycYeastCycle}
\title{cycYeastCycle}
\format{A dataframe containing 12 columns(column 1 = transcript name,
  column 2 to 12 = time points from 2 minutes to 162 minutes after
  recovery phase).}
\source{
Orlando D. A., et al. (2008). Global control of cell-cycle
  transcription by coupled CDK and network oscillators. \emph{Nature},
  \bold{453(7197)}, 944--947.
}
\usage{
cycYeastCycle
}
\description{
This data set lists expression profiles of 20 cycling
  transcripts with 16-minutes resolution covering about two yeast cell
  cycles.
}

