% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MetaCycle}
\alias{MetaCycle}
\alias{MetaCycle-package}
\title{MetaCycle
An integrated R package to evaluate periodicity in large scale data}
\description{
MetaCycle
An integrated R package to evaluate periodicity in large scale data
}
\author{
Gang Wu \email{wggucas@gmail.com},
Ron Anafi \email{ranafi@gmail.com},
Xavier Li \email{haohan.li.97@gmail.com},
Matthew Carlucci \email{matt.carlucci.96@gmail.com},
John Hogenesch \email{hogenesch@gmail.com},
Michael Hughes \email{michael.evan.hughes@gmail.com},
Karl Kornacker \email{kornacker@midohio.twcbc.com}
Maintainer: Gang Wu \email{wggucas@gmail.com}
}
